requirements.txt
Flask==2.3.3
firebase-admin==7.1.0
google-generativeai==0.7.1
numpy==2.3.3
Pillow==11.0.0
python-dotenv==1.0.1
requests==2.31.0
tensorflow==2.20.0